use exit code
