+++ pyproject.toml
@@ -66,7 +66,7 @@ style = [
 ]
 
 [project.scripts]
-bidspm = "bidspm.bidspm:cli"
+bidspm = "bidspm.cli:cli"
 validate_model = "bidspm.validate:cli"
 
 [project.urls]
