+++ src/bidspm/data/allowed_actions.json
@@ -1,0 +1,14 @@
+[
+  "bms",
+  "bms-bms",
+  "bms-posterior",
+  "copy",
+  "contrasts",
+  "create_roi",
+  "default_model",
+  "preprocess",
+  "results",
+  "smooth",
+  "specify_only",
+  "stats"
+]
