+++ src/cli/getOptionsFromCliArgument.m
@@ -133,23 +133,6 @@ function opt = optionsStats(opt, args, action)
   end
 end
 
-function value = bidsAppsActions()
-
-  value = {'copy'; ...
-           'create_roi'; ...
-           'preprocess'; ...
-           'smooth'; ...
-           'default_model'; ...
-           'stats'; ...
-           'contrasts'; ...
-           'results'; ...
-           'specify_only'; ...
-           'bms'; ...
-           'bms-posterior'; ...
-           'bms-bms' ...
-          };
-end
-
 function opt = getOptions(args)
 
   if isstruct(args.Results.options)
